backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
requests==2.32.5
pydantic==2.11.7
pypdf==5.9.0
python-multipart==0.0.20